slurmd: add support for cgroups in slurm_NODENAME
When slurmd is build with multiple slurmd per node support, the
name of the subdirectory where the cgroups are created changes to
include the hostname (cpuset/slurm_NODENAME instead of
cpuset/slurm). This patch supports this by trying both and add an
option to select a node name if the nodename is not the local
hostname.