text protocol: Add support for meta_data in PUTVAL
This adds the support for string meta_data in PUTVAL. This allows users
to send meta_data using either the unixsock plugin or the exec plugin.

The meta_data is specified using options of the form meta:<KEY>=<VALUE>.
As an example, meta:foo=bar will add an entry with key "foo" and value
"bar".