Fix most truncation warnings from GCC 8.2 (WIP)
GCC 8.2 now warns about possible truncation in snprintf() and
strncat(). As for snprintf(), we introduce an helper ssnprintf() to
handle this correctly. It silently adds the missing '\0' at the end.
The change is mostly mechanical (some help from Coccinelle) and may be
a bit overzealous as in some case, it may be unneeded (GCC doesn't
detect all cases either).

vsnprintf() is also handled by putting the terminal null character in
case of overflow.

As for strncat(), the situation is far more grim. GCC seems to
consider safe truncation should not happen. Therefore, when it is able
to infer the buffer length, like with `strncat(out, src, sizeof(out) -
strlen(out) - 1)`, it will trigger an error. Therefore, when this
happens, we use ssnprintf() instead which is strictly equivalent but
silence the truncation by capturing the return value. An alternative
would be to introduce `strlcat()`.