configure.ac: run dpdk build tests only if pkgconfig fails
The AC_CHECK_LIB test runs unconditionally, and fails with DPDK 18.11
when built with Meson as there is no longer a libdpdk.so linker script,
but only a pkg-config file, so -ldpdk (which is what AC_CHECK_LIB runs)
fails.

Use AC_LINK_IFELSE instead, with compiler and linker flags set
appropriately.

Signed-off-by: Luca Boccassi <bluca@debian.org>