lvm plugin: turn deprecation error into a warning
When you include lmv2app.h it warns that this is deprecated and you have
to use D-BUS.

This warning turns into an error since we build with -Werror by default.
We know using the lvm2 lib is deprecated but until someone writes a
replacement using D-BUS it's all we have.

So turn the error into a warning.