plugin: processes: use a const ref string and strlen i.s.o. magic offset
The simple parser is based on 'strstr', whereas the actual data does
include a white-space character to separate the 'id' (procs_running)
from the 'value' (a number). Yet parsing the number can off course only
take place 'after' the white-space. Subsequently, not finding a
'white-space' should be treated as an error condition.

Adding the 'white-space' in const string and defining the offset based
on 'strlen' fixes this properly.