This plugin, named mdevents, is responsible for gathering the events from RAID arrays that were written to syslog by mdadm utility (which is a user-space software for managing the RAIDs). Then, based on configuration provided by user, plugin will decide whether to send the collectd notification or not.
Mdevents needs the syslog and mdadm to be present on a platform that collectd is launched.

Based on the naming of existing dpdk plugins (dpdkstat for metrics, dpdkevents for notifications) and the fact that there is already plugin called md for gathering metrics from RAIDs, giving this plugin name mdevents felt like the best option.

A slight change to ignorelist API was introduced - it is now possible to retrieve the count of nodes in array and the current ignore setting for given array.

ChangeLog: New plugin to read RAID events