amqp1: Add options to limit send queue length
ChangeLog: amqp1 plugin: Add options to limit send queue length.

Add two options to control send queue length and behavior:
 - SendQueueLimit and SendQueueDropNew

If SendQueueLimit is set to an integer greater than 0, the send queue in the
amqp1 plugin will be constrained to that number of entries to avoid unbounded
memory use in the case that there is no amqp1 connection. If the
SendQueueDropNew parameter is set to true, new messages will be dropped when
the queue is full and a new message is enqueued, otherwise the oldest message
in the queue will be dropped. The default is for the oldest message to be
dropped.

Signed-off-by: Ryan McCabe <rmccabe@redhat.com>