amqp1: Add option to limit internal send queue length
ChangeLog: amqp1 plugin: Add option to limit internal send queue length.

Add a config  option to control the internal plugin send queue
length: SendQueueLimit

If SendQueueLimit is set to an integer greater than 0, the send queue
in the amqp1 plugin will be constrained to that number of entries to
avoid unbounded memory use in the case that there is no amqp1
connection. If a new message is queued when the queue has reached
maximum length, the oldest message in the queue will be discarded.

Signed-off-by: Ryan McCabe <rmccabe@redhat.com>