src/daemon/metric.[ch]: Data types and functions representing metrics.
This implementation closely mirrors the Prometheus / OpenMetrics
protocol buffer.

A "metric family" is a set of metrics, all with the same name (but
different labels and/or label values). The identity is split between the
two types: the "metric family" holds the "name" while the "metric" holds
the labels. Likewise, the "metric family" holds the metric type, the
"metric" holds the metric value.

This commit contains all required changes to compile the daemon, but
pretty much everything else still fails to build. The tests have not yet
updated either.

The metric_marshal_text() function uses a strbuf_t* and generates the
OpenMetric exposition format.