capabilities: conditionalize mhd callback return type
libmicrohttpd introduced an API change with respect its callback
return types. Now its signature says that an enum must be returned
instead of an int. This causes collectd to not finish compilation
because of the compiler warnings this arised (in addition to -Werror)

Add conditional compiling so systems with microhttpd >=0.9.71 can
compile collectd with -Werror

Fixes #3511