mqtt plugin: encapsulate mosquitto_loop
Encapsulate the conditional compilation in a helper function to clean up
the existing code and to avoid duplicating it into new locations with
future changes.

The other option would have been to remove the support for the major
version 0, which would have allowed removing the conditional
compilation. The conditional compilation was added in the commit
416aab49a9ae19b993b747b5fca92fc3501fff5d ("mqtt plugin: Add support for
libmosquitto, major version >0.") in 2015. Based on the Mosquitto Git
history Mosquitto 1.x has existed at least since 2014, and Mosquitto
recently released v2, but it is not known if collectd still needs to
support Mosquitto 0.x.

Signed-off-by: Hannu Lounento <hannu.lounento@iki.fi>