Feature-redfish: enabling libredfish logging on collectd debug mode
Changelog: enabling libredfish logging on collectd debug mode

In order for `libredfish` to log internal information, which are more
than useful to debugging purposes, a printing function has to be
registered through a hook : a global function pointer named `gDebugFunc`.

The `plugin_log` function is registered as the logging function through
the aforementioned hook by the `redfish_init` function, when the
`--enable-debug` option is specified to the `configure` script.