Fix CI failures caused by unsupported distros and updates to dependencies (#3975)
* [ci][gha] Replace trusy with Bionic and Focal

Ubuntu 14.04 (Trusty) is out of standard support [1].
``make check`` fails for test_capabilities, as noted in [2].
[3] indicates that the cause is glibc, but that updates are not expected
to the version in trusty.

This PR replaces trusty with Ubuntu 18.04 (Bionic) and 20.04 (Focal).

[1] https://wiki.ubuntu.com/Releases
[2] #3936
[3] #3927 (comment)

* [ci][cirrus] Make Valgrind error on defininte memory leaks only

Valgrind gives errors when it find possible leaks;
update the options to only error on definite leaks.

This is done using the VALGRIND_OPTS env var, which is used by valgrind
when it is invoked.

* [ci][gha] Make Valgrind error on defininte memory leaks only

Valgrind gives errors when it find possible leaks;
update the options to only error on definite leaks.

This is done using the VALGRIND_OPTS env var, which is used by valgrind
when it is invoked.