smart: check udev_enumerate_scan_devices() return value and unify log messages (#3984)
* Check udev_enumerate_scan_devices return value

This change checks the return value of the function and cancels the call
if the returned integer is not greater than or equal to 0.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>

* Unify and fix log messages

The error log messages were not consistent and had no prefix. This
commit adds the uniform prefix 'smart plugin:' to each log message. While
we're at it, I also removed the punctuation mark at the end of the
sentences.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>