[collectd 6] exec: make PUTMETRIC work
The cmd_handle_putmetric function checks if the command actually is
a PUTMETRIC command, at least that is what is supposed to check.
Prior to this fix it actually checked for PUTVAL and always prints a

    -1 Unexpected command: `PUTMETRIC'.

error. While at it also remove the development printf that results in

    set_option("type", "GAUGE")
    set_option("type", "Counter")
    …

being printed during normal operation.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>