src/daemon/plugin.c: split write thread stop and free in plugin_shutdown_all()
This fixes an issue with write plugins that assume callback teardown to happen
in a specific order (e.g. those that use the "set the free_function to NULL"
trick mentioned in the comment close to this change to prevent a double-free
of the user_data).

I found at least the following plugins using this trick:

 - write_http
 - write_riemann
 - write_stackdriver
 - write_syslog
 - write_tsdb

This should fix use-after-frees during teardown of these plugins by restoring
the freeing order to what it was before
Commit 55efb56a ("src/daemon/plugin.c: Use one thread per write plugin").

Fixes: 55efb56a ("src/daemon/plugin.c: Use one thread per write plugin")
Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>