smart: workaround NVMe API buffer overruns on Linux>=5.19-rc1
Around Linux 5.19-rc1 NVMe ioctl behavior changed, and it now can write
past data_len bytes when buffers are dword-aligned. It's highly likely for
a target buffer to be aligned at four-byte boundaries, causing neighboring
data corruption.

The workaround is to ask NVMe ioctls to write to artificially shifted
buffers. That triggers slower path in the kernel, where it asks NVMe
drive to write to a kernel buffer first, and then copies data back to the
user space. That path happens to honor user buffer boundaries.