swap plugin: Use different metric families for I/O reporting in bytes.
Depending on the "ReportIO" setting, the plugin may emit paging operations
or bytes. Use a different metric family for these two metrics.

This also moves the decision whether or not to report I/O to the
`swap_submit_io`. This fixes a bug on NetBSD where I/O was reported
even if "ReportIO" was set to false.