src/daemon/metric.h: Add the `VALUE_T()` generic macro.
`VALUE_T(x)` is a generic macro that initializes and returns a `value_t`.
Depending on the type of the argument `x`, it will either initialize the
`value_t` as a `counter_t`, `derive_t`, or `gauge_t`.

This means that the previous expression:

    value_t x = (value_t){.derive = y};

can now be written as:

    value_t x = VALUE_T(y);

Passing in an incompatible type, such as an `int`, will cause a compile time
error:

```
src/cpu.c: In function 'module_register':
./src/daemon/metric.h:56:12: error: '_Generic' selector of type 'int' is not compatible with any association
   56 |   _Generic((x),                                                                \
      |            ^
src/cpu.c:929:17: note: in expansion of macro 'VALUE_T'
  929 |   value_t foo = VALUE_T((int){42});
      |                 ^~~~~~~
```