ci: Remove the "statgrab" task from CirrusCI.
The "statgrab" job builds collectd with the "statgrab" library and then expects
certain plugins to be linked against it. However, since the build runs on
Linux, the plugins are actually using the "native" implementation. The linker,
noticing that the plugin is not referring to any symbol in the statgrab
library, does not actually produce a link, causing the test to fail.

The native implementations provide much richer metrics than the libstatgrab
based implementations, so we'll drop the test.