write_prometheus plugin: Store resource attributes with the metric, not the metric family.
For Prometheus output, the plugin groups all metrics with the same name into
one `metric_family_t`. This caused problems when collectd handled metrics from
multiple resources.

To solve this issue, we're somewhat abusing the data structure and store
per-metric resource attributes in the `family` field. That means for the
metrics stored in the *write_prometheus plugin* `(metric_t).family` does not
point back to the metric family containing the metric.

Fixes: #4283