fix(lua): Use `lua_tonumber` to support 32-bit platforms
Previously, `luaC_tovalue` used `lua_tointeger` for variables of type
"derive", "counter", and "absolute". However, on 32-bit platforms,
`lua_Integer` may be only 32 bits wide [1], so if you set `values` to a
value larger than 2^31-1 from Lua, the value of `lua_tointeger` will be
unspecified [2].

This commit changes the code to use `lua_tonumber` on 32-bit platforms,
which returns a float, which will then be cast to the (typedef'ed to
`[u]int64_t`) integer types `derive_t`, `counter_t`, and `absolute_t`,
ensuring that large values are not truncated.

When `lua_Integer` is 64 bits wide, the code will still use
`lua_tointeger` to ensure that integers between 2^53 and 2^63-1 do not
lose precision on newer versions of Lua with dedicated integer
types [3].

[1]: https://www.lua.org/manual/5.1/manual.html#lua_Integer

[2]: https://www.lua.org/manual/5.1/manual.html#lua_tointeger

[3]: https://www.lua.org/manual/5.3/manual.html#8.1